Add unit tests for single value generator
